rebasing, picking, and squashing to clean up revision history.
lots of changes to memory management all over the vq code space. for the most part, this is using {delete []} in place of {delete} for arrays and also the proper application of free() for malloc() arrays... or at least it makes valgrind happy. still seeing what is probably the major problem, which appears to be some sort of misdirection on the stdout (or something like this). vargrind shows unallocated pointer(s) when QuakeLibIO.cpp:quakelib::ModelSweeps::write_ascii() runs; not sure if it is the value that is not properly allocated (aka, the _normal_final variable) or the file stream (std::ostream &out_stream)

tracking down heisenbug; looks like it is isolated to one location in GreensINit.cpp